+++ R/step-first.R
@@ -28,6 +28,21 @@
 #' @examples
 #' library(dplyr, warn.conflicts = FALSE)
 #'
+#' # If you have a data.table, using it with any dplyr generic will
+#' # automatically convert it to a lazy_dt object
+#' dt <- data.table::data.table(x = 1:10, y = 10:1)
+#' dt %>% filter(x == y)
+#' dt %>% mutate(z = x + y)
+#'
+#' # Note that dtplyr will avoid mutating the input data.table, so the
+#' # previous translation includes an automatic copy(). You can avoid this
+#' # with a manual call to lazy_dt()
+#' dt %>%
+#'   lazy_dt(immutable = FALSE) %>%
+#'   mutate(z = x + y)
+#'
+#' # If you have a data frame, you can use lazy_dt() to convert it to
+#' # a data.table:
 #' mtcars2 <- lazy_dt(mtcars)
 #' mtcars2
 #' mtcars2 %>% select(mpg:cyl)
@@ -38,10 +53,13 @@
 #' mtcars2 %>% transmute(cyl2 = cyl * 2, vs2 = vs * 2)
 #' mtcars2 %>% filter(cyl == 8) %>% mutate(cyl2 = cyl * 2)
 #'
+#' # Learn more about translation in vignette("translation")
 #' by_cyl <- mtcars2 %>% group_by(cyl)
 #' by_cyl %>% summarise(mpg = mean(mpg))
 #' by_cyl %>% mutate(mpg = mean(mpg))
-#' by_cyl %>% filter(mpg < mean(mpg)) %>% summarise(hp = mean(hp))
+#' by_cyl %>%
+#'   filter(mpg < mean(mpg)) %>%
+#'   summarise(hp = mean(hp))
 lazy_dt <- function(x, name = NULL, immutable = TRUE, key_by = NULL) {
   if (!is.data.table(x)) {
     if (!immutable) {