+++ R/step-join.R
@@ -94,6 +94,12 @@ left_join.dtplyr_step <- function(x, y, ..., by = NULL, copy = FALSE, suffix = c
   }
 }
 
+#' @export
+left_join.data.table <- function(x, y, ...) {
+  x <- lazy_dt(x)
+  left_join(x, y, ...)
+}
+
 #' @importFrom dplyr right_join
 #' @export
 right_join.dtplyr_step <- function(x, y, ..., by = NULL, copy = FALSE, suffix = c(".x", ".y")) {
@@ -106,6 +112,13 @@ right_join.dtplyr_step <- function(x, y, ..., by = NULL, copy = FALSE, suffix = 
     step_join(y, x, on = by, style = "left", suffix = suffix)
   }
 }
+
+#' @export
+right_join.data.table <- function(x, y, ...) {
+  x <- lazy_dt(x)
+  right_join(x, y, ...)
+}
+
 
 step_subset_on <- function(x, y, i, on) {
   step_subset(x,
@@ -125,6 +138,12 @@ inner_join.dtplyr_step <- function(x, y, ..., by = NULL, copy = FALSE, suffix = 
   step_join(x, y, on = by, style = "inner", suffix = suffix)
 }
 
+#' @export
+inner_join.data.table <- function(x, y, ...) {
+  x <- lazy_dt(x)
+  inner_join(x, y, ...)
+}
+
 #' @importFrom dplyr full_join
 #' @export
 full_join.dtplyr_step <- function(x, y, ..., by = NULL, copy = FALSE, suffix = c(".x", ".y")) {
@@ -132,6 +151,12 @@ full_join.dtplyr_step <- function(x, y, ..., by = NULL, copy = FALSE, suffix = c
   by <- dtplyr_common_by(by, x, y)
 
   step_join(x, y, on = by, style = "full", suffix = suffix)
+}
+
+#' @export
+full_join.data.table <- function(x, y, ...) {
+  x <- lazy_dt(x)
+  full_join(x, y, ...)
 }
 
 #' @importFrom dplyr anti_join
@@ -143,6 +168,12 @@ anti_join.dtplyr_step <- function(x, y, ..., by = NULL, copy = FALSE) {
   step_join(x, y, on = by, style = "anti")
 }
 
+#' @export
+anti_join.data.table <- function(x, y, ...) {
+  x <- lazy_dt(x)
+  anti_join(x, y, ...)
+}
+
 #' @importFrom dplyr semi_join
 #' @export
 semi_join.dtplyr_step <- function(x, y, ..., by = NULL, copy = FALSE) {
@@ -150,6 +181,12 @@ semi_join.dtplyr_step <- function(x, y, ..., by = NULL, copy = FALSE) {
   by <- dtplyr_common_by(by, x, y)
 
   step_join(x, y, on = by, style = "semi")
+}
+
+#' @export
+semi_join.data.table <- function(x, y, ...) {
+  x <- lazy_dt(x)
+  semi_join(x, y, ...)
 }
 
 # helpers -----------------------------------------------------------------