+++ R/step-mutate.R
@@ -64,6 +64,12 @@ mutate.dtplyr_step <- function(.data, ...) {
   step_mutate(.data, dots, nested)
 }
 
+#' @export
+mutate.data.table <- function(.data, ...) {
+  .data <- lazy_dt(.data)
+  mutate(.data, ...)
+}
+
 nested_vars <- function(.data, dots, all_vars) {
   new_vars <- character()
   all_new_vars <- unique(names(dots))